Separate BillingHandler integration and unit tests
